+++ comics/comics/forbetterorforworse.py
@@ -16,4 +16,4 @@ class Crawler(GoComicsComCrawlerBase):
     time_zone = 'US/Eastern'
 
     def crawl(self, pub_date):
-        return self.crawl_helper('For Better or For Worse', pub_date)
+        return self.crawl_helper('forbetterorforworse', pub_date)