+++ comics/comics/theboondocks.py
@@ -5,7 +5,7 @@ from comics.core.comic_data import ComicDataBase
 class ComicData(ComicDataBase):
     name = 'The Boondocks'
     language = 'en'
-    url = 'http://www.gocomics.com/theboondocks'
+    url = 'http://www.gocomics.com/boondocks'
     start_date = '1999-04-19'
     rights = 'Aaron McGruder'
 
@@ -16,4 +16,4 @@ class Crawler(GoComicsComCrawlerBase):
     time_zone = 'US/Mountain'
 
     def crawl(self, pub_date):
-        return self.crawl_helper('The Boondocks', pub_date, 'boondocks')
+        return self.crawl_helper('boondocks', pub_date)